+++ CHANGELOG.md
@@ -1,12 +1,18 @@
 # Getdown Releases
 
+## 1.8.7 - Unreleased
+
+* Reinstated env var support in `appbase` property.
+
+* Fixed issue with `myIpAddress()` in PAC proxy support.
+
 ## 1.8.6 - June 4, 2019
 
 * Fixed issues with PAC proxy support: added `myIpAddress()`, fixed `dnsResolve()`, fixed crash
   when detecting PAC proxy.
 
-* Reverted env var support in `appbase` and `latest` properties. It's causing problems that need to
-  be investigated.
+* Reverted env var support in `appbase` property. It's causing problems that need to be
+  investigated.
 
 ## 1.8.5 - May 29, 2019